# openApi.yaml
openapi: 3.0.3
info:
  title: Commulab System API
  description: >
    API untuk autentikasi, manajemen pasien, sesi konseling,
    chat AI, transkrip, evaluasi, dan laporan statistik.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /patients:
    get:
      tags: [Patients]
      summary: Ambil semua pasien
      description: >
        Mengambil daftar pasien berdasarkan user yang login. 
        User biasa hanya bisa melihat pasien miliknya sendiri dan pasien global.
        Admin bisa melihat semua pasien.
      responses:
        "200":
          description: Daftar pasien berhasil diambil
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    patient_id:
                      type: integer
                      example: 1
                    patient_name:
                      type: string
                      example: "Budi Santoso"
                    personality_traits:
                      type: object
                      nullable: true
                      example: { "anxiety_level": "high", "openness": "medium" }
                    avatar_path:
                      type: string
                      example: "/models/default.glb"
                    created_at:
                      type: string
                      format: date-time
                      example: "2024-12-27T10:00:00.000Z"
                    is_active:
                      type: boolean
                      example: true
                    is_global:
                      type: boolean
                      example: false
                    user_id:
                      type: integer
                      example: 1
              examples:
                success:
                  summary: Daftar pasien
                  value:
                    - patient_id: 1
                      patient_name: "Budi Santoso"
                      personality_traits: { "anxiety_level": "high" }
                      avatar_path: "/models/default.glb"
                      created_at: "2024-12-27T10:00:00.000Z"
                      is_active: true
                      is_global: false
                      user_id: 1
                    - patient_id: 2
                      patient_name: "Siti Aminah"
                      personality_traits: { "depression_level": "medium" }
                      avatar_path: "/models/default.glb"
                      created_at: "2024-12-26T09:00:00.000Z"
                      is_active: true
                      is_global: true
                      user_id: 2
        "500":
          description: Error server
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string

    post:
      tags: [Patients]
      summary: Tambah pasien baru
      description: >
        Membuat pasien baru dengan data lengkap. Sistem akan otomatis 
        generate embeddings (RAG) dari data pasien untuk meningkatkan 
        akurasi response AI. Jika embedding gagal, pasien tetap dibuat 
        tanpa RAG data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patient_name, background_story]
              properties:
                patient_name:
                  type: string
                  description: Nama lengkap pasien
                  example: "Budi Santoso"
                background_story:
                  type: string
                  description: Latar belakang dan riwayat pasien
                  example: "Mahasiswa semester 6 yang mengalami anxiety sejak pandemi..."
                personality_type:
                  type: string
                  nullable: true
                  description: Tipe kepribadian (opsional)
                  example: "introvert"
                symptom_intensity:
                  type: integer
                  minimum: 1
                  maximum: 10
                  nullable: true
                  description: Intensitas gejala 1-10 (opsional)
                  example: 7
                age:
                  type: integer
                  nullable: true
                  description: Usia pasien (harus positif)
                  example: 22
                gender:
                  type: string
                  nullable: true
                  description: Jenis kelamin
                  example: "Laki-laki"
                occupation:
                  type: string
                  nullable: true
                  description: Pekerjaan/profesi
                  example: "Mahasiswa"
                marital_status:
                  type: string
                  nullable: true
                  description: Status pernikahan
                  example: "Belum Menikah"
                personality_traits:
                  type: object
                  nullable: true
                  description: Traits kepribadian dalam bentuk key-value
                  example: { "anxiety_level": "high", "social_comfort": "low" }
                is_global:
                  type: boolean
                  nullable: true
                  description: Apakah pasien bisa diakses semua user
                  example: false
                  default: false
            examples:
              complete:
                summary: Data lengkap
                value:
                  patient_name: "Budi Santoso"
                  background_story: "Mahasiswa yang mengalami anxiety disorder"
                  personality_type: "introvert"
                  symptom_intensity: 7
                  age: 22
                  gender: "Laki-laki"
                  occupation: "Mahasiswa"
                  marital_status: "Belum Menikah"
                  personality_traits:
                    anxiety_level: "high"
                    social_comfort: "low"
                    openness: "medium"
                  is_global: false
              minimal:
                summary: Data minimal (required only)
                value:
                  patient_name: "Siti Aminah"
                  background_story: "Pasien dengan riwayat depresi ringan"
      responses:
        "201":
          description: Pasien berhasil ditambahkan
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "Pasien berhasil ditambahkan"
                  data:
                    type: object
                    properties:
                      patient_id:
                        type: integer
                        example: 1
                      user_id:
                        type: integer
                        example: 1
                      patient_name:
                        type: string
                        example: "Budi Santoso"
                      background_story:
                        type: string
                      personality_type:
                        type: string
                        nullable: true
                      symptom_intensity:
                        type: integer
                        nullable: true
                      age:
                        type: integer
                        nullable: true
                      gender:
                        type: string
                        nullable: true
                      occupation:
                        type: string
                        nullable: true
                      marital_status:
                        type: string
                        nullable: true
                      personality_traits:
                        type: object
                        nullable: true
                      avatar_path:
                        type: string
                        example: "/models/default.glb"
                      is_active:
                        type: boolean
                        example: true
                      is_global:
                        type: boolean
                        example: false
                      knowledge_base:
                        type: array
                        nullable: true
                        description: RAG embeddings (jika berhasil generate)
                        items:
                          type: object
                          properties:
                            text:
                              type: string
                            vector:
                              type: array
                              items:
                                type: number
              examples:
                success:
                  summary: Pasien berhasil dibuat
                  value:
                    status: success
                    message: "Pasien berhasil ditambahkan"
                    data:
                      patient_id: 1
                      user_id: 1
                      patient_name: "Budi Santoso"
                      background_story: "Mahasiswa yang mengalami anxiety"
                      personality_type: "introvert"
                      symptom_intensity: 7
                      age: 22
                      gender: "Laki-laki"
                      occupation: "Mahasiswa"
                      marital_status: "Belum Menikah"
                      personality_traits:
                        anxiety_level: "high"
                      avatar_path: "/models/default.glb"
                      is_active: true
                      is_global: false
                      knowledge_base: [
                        {
                          text: "Mahasiswa semester 6 yang mengalami anxiety",
                          vector: [0.123, 0.456, ...]
                        }
                      ]
        "400":
          description: Validasi gagal
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
              examples:
                missing_required:
                  summary: Field required tidak diisi
                  value:
                    status: error
                    message: "patient_name dan background_story harus diisi"
                invalid_intensity:
                  summary: Intensitas tidak valid
                  value:
                    status: error
                    message: "symptom_intensity harus antara 1-10"
                invalid_age:
                  summary: Age tidak valid
                  value:
                    status: error
                    message: "age harus bernilai positif"
        "500":
          description: Error server
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
  /patients/{id}:
    get:
      tags: [Patients]
      summary: Ambil detail pasien
      description: >
        Mengambil detail lengkap pasien berdasarkan ID.
        User biasa hanya bisa mengakses pasien miliknya sendiri atau pasien global.
        Admin bisa mengakses semua pasien.
      parameters:
        - name: id
          in: path
          required: true
          description: ID pasien
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Detail pasien
          content:
            application/json:
              schema:
                type: object
                properties:
                  patient_id:
                    type: integer
                  user_id:
                    type: integer
                  patient_name:
                    type: string
                  background_story:
                    type: string
                  personality_type:
                    type: string
                    nullable: true
                  symptom_intensity:
                    type: integer
                    nullable: true
                  age:
                    type: integer
                    nullable: true
                  gender:
                    type: string
                    nullable: true
                  occupation:
                    type: string
                    nullable: true
                  marital_status:
                    type: string
                    nullable: true
                  personality_traits:
                    type: object
                    nullable: true
                  avatar_path:
                    type: string
                  profile_image:
                    type: string
                    nullable: true
                  is_active:
                    type: boolean
                  is_global:
                    type: boolean
                  knowledge_base:
                    type: array
                    nullable: true
                  created_at:
                    type: string
                    format: date-time
              examples:
                success:
                  summary: Detail pasien
                  value:
                    patient_id: 1
                    user_id: 1
                    patient_name: "Budi Santoso"
                    background_story: "Mahasiswa yang mengalami anxiety"
                    personality_type: "introvert"
                    symptom_intensity: 7
                    age: 22
                    gender: "Laki-laki"
                    occupation: "Mahasiswa"
                    marital_status: "Belum Menikah"
                    personality_traits:
                      anxiety_level: "high"
                    avatar_path: "/models/default.glb"
                    profile_image: null
                    is_active: true
                    is_global: false
                    knowledge_base: []
                    created_at: "2024-12-27T10:00:00.000Z"
        "401":
          description: User tidak terautentikasi
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: "User not authenticated"
        "404":
          description: Pasien tidak ditemukan
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: "Pasien tidak ditemukan"
        "500":
          description: Error server
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
  /patients/model/{patientId}:
    get:
      tags: [Patients]
      summary: Ambil path model avatar pasien
      description: >
        Mengambil path file 3D model (.glb) untuk avatar pasien.
        Jika pasien tidak memiliki avatar custom, akan return default avatar.
      parameters:
        - name: patientId
          in: path
          required: true
          description: ID pasien
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Path avatar pasien
          content:
            application/json:
              schema:
                type: object
                properties:
                  avatar_path:
                    type: string
                    description: Path ke file 3D model avatar
                    example: "/models/default.glb"
              examples:
                custom_avatar:
                  summary: Avatar custom
                  value:
                    avatar_path: "/models/patient_1_avatar.glb"
                default_avatar:
                  summary: Avatar default
                  value:
                    avatar_path: "/models/default.glb"
        "404":
          description: Pasien tidak ditemukan
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Pasien tidak ditemukan"
        "500":
          description: Error server
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /patients/migrate-embeddings:
    post:
      tags: [Patients]
      summary: Migrasi embedding pasien (RAG)
      description: >
        Melakukan migrasi/regenerasi embeddings untuk semua pasien 
        yang belum memiliki knowledge_base atau perlu update.
        Proses ini menggunakan Google Gemini API untuk generate 
        vector embeddings dari data pasien. Endpoint ini untuk 
        maintenance/migration purposes.
      responses:
        "200":
          description: Migrasi embedding berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "Berhasil migrasi 5 dari 10 pasien."
              examples:
                success:
                  summary: Migrasi berhasil
                  value:
                    status: success
                    message: "Berhasil migrasi 5 dari 10 pasien."
                no_patients:
                  summary: Tidak ada pasien yang perlu migrasi
                  value:
                    status: success
                    message: "Berhasil migrasi 0 dari 0 pasien."
        "500":
          description: Error saat migrasi
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to generate embeddings"
...
  /reports/patient/{patient_id}:
    get:
      tags: [Reports]
      summary: Statistik per pasien
      description: >
        Mengambil statistik lengkap untuk pasien tertentu, termasuk
        total sessions, completed sessions, average scores, dan detail
        setiap session dengan evaluasinya.
      parameters:
        - name: patient_id
          in: path
          required: true
          description: ID pasien
          schema:
            type: integer
          example: 5
      responses:
        "200":
          description: Statistik pasien berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  patient_id:
                    type: integer
                    example: 5
                  totalSessions:
                    type: integer
                    description: Total semua session
                    example: 8
                  completedSessions:
                    type: integer
                    description: Total session yang selesai
                    example: 6
                  evaluatedSessions:
                    type: integer
                    description: Total session yang sudah dievaluasi
                    example: 5
                  averageScores:
                    type: object
                    properties:
                      empathy:
                        type: integer
                        example: 85
                      question:
                        type: integer
                        example: 78
                      ethics:
                        type: integer
                        example: 0
                      overall:
                        type: integer
                        example: 54
                  sessions:
                    type: array
                    items:
                      type: object
                      properties:
                        session_id:
                          type: integer
                        start_time:
                          type: string
                          format: date-time
                        end_time:
                          type: string
                          format: date-time
                          nullable: true
                        status:
                          type: string
                        hasEvaluation:
                          type: boolean
                        scores:
                          type: object
                          nullable: true
                          properties:
                            empathy_score:
                              type: integer
                            question_score:
                              type: integer
              examples:
                success:
                  summary: Statistik pasien
                  value:
                    patient_id: 5
                    totalSessions: 8
                    completedSessions: 6
                    evaluatedSessions: 5
                    averageScores:
                      empathy: 85
                      question: 78
                      ethics: 0
                      overall: 54
                    sessions:
                      - session_id: 1
                        start_time: "2024-12-20T10:00:00.000Z"
                        end_time: "2024-12-20T11:00:00.000Z"
                        status: "finished"
                        hasEvaluation: true
                        scores:
                          empathy_score: 85
                          question_score: 80
                      - session_id: 2
                        start_time: "2024-12-21T14:00:00.000Z"
                        end_time: null
                        status: "ongoing"
                        hasEvaluation: false
                        scores: null
        "500":
          description: Error server
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...